# api-docs/api-docs_v1.1.json
{"openapi":"3.1.0","info":{"title":"CodeGuard API","description":"API для двухфакторной аутентификации и управления OTP кодами","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Аутентификация","description":"Регистрация и авторизация пользователей"},{"name":"OTP коды","description":"Генерация и валидация OTP кодов"},{"name":"Пользователи","description":"Управление пользователями и их настройками"},{"name":"Конфигурация OTP","description":"Управление настройками OTP кодов"}],"paths":{"/api/otp/generate":{"post":{"tags":["OTP коды"],"summary":"Сгенерировать OTP код","description":"Создает OTP код и отправляет его на указанные каналы. Требуется авторизация","operationId":"generateOtpCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationDTO"}}},"required":true},"responses":{"200":{"description":"Код отправлен на все каналы","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpGenerationResponse"},"examples":{"SuccessAllChannels":{"summary":"Успешная отправка на все каналы","description":"SuccessAllChannels","value":{"code":"123456","status":"SENT","successfulChannels":["SMS","EMAIL","TELEGRAM"],"failedChannels":{},"expiresAt":"2023-12-31T23:59:59"}}}}}},"207":{"description":"Код отправлен на часть каналов","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpGenerationResponse"},"examples":{"PartialSuccess":{"summary":"Частичная отправка","description":"PartialSuccess","value":{"code":"654321","status":"PARTIAL","successfulChannels":["SMS","TELEGRAM"],"failedChannels":{"EMAIL":"SMTP server unavailable"},"expiresAt":"2023-12-31T23:59:59"}}}}}},"503":{"description":"Ошибка отправки на все каналы","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpGenerationResponse"},"examples":{"FailedAllChannels":{"summary":"Ошибка отправки на все каналы","description":"FailedAllChannels","value":{"code":"987654","status":"FAILED","successfulChannels":[],"failedChannels":{"SMS":"Invalid phone number","EMAIL":"SMTP server unavailable","TELEGRAM":"Bot blocked by user"},"expiresAt":"2023-12-31T23:59:59"}}}}}},"403":{"description":"Требуется авторизация"}}}},"/api/auth/register":{"post":{"tags":["Аутентификация"],"summary":"Регистрация пользователя","description":"Создает нового пользователя в системе","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationDTO"}}},"required":true},"responses":{"200":{"description":"Пользователь успешно зарегистрирован","content":{"*/*":{"schema":{"type":"string","example":"Пользователь сохранен"}}}},"400":{"description":"Некорректные данные пользователя","content":{"*/*":{"schema":{"type":"string","example":{"username":"Username must be between 3 and 50 characters"}}}}}}}},"/api/auth/login":{"post":{"tags":["Аутентификация"],"summary":"Авторизация пользователя","description":"Аутентифицирует пользователя и возвращает JWT токен","operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationDTO"}}},"required":true},"responses":{"200":{"description":"Успешная авторизация","content":{"*/*":{"schema":{"type":"string","example":{"jwt-token":"string"}}}}},"401":{"description":"Неверные учетные данные","content":{"*/*":{"schema":{"type":"string","example":{"message":"Incorrect credentials!"}}}}}}}},"/api/user/setMessageTemplate":{"patch":{"tags":["Пользователи"],"summary":"Установить шаблон сообщения","description":"Обновляет шаблон сообщения для OTP кода. Требуется авторизация","operationId":"setTemplateMessage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageTemplateDTO"}}},"required":true},"responses":{"200":{"description":"Шаблон успешно обновлен","content":{"*/*":{"schema":{"type":"string","example":"Шаблон обновлен"}}}},"400":{"description":"Некорректный шаблон","content":{"*/*":{"schema":{"type":"string","example":{"template":"Template must contain '{code}' substring"}}}}},"403":{"description":"Доступ запрещен"}}}},"/api/otp/config":{"get":{"tags":["Конфигурация OTP"],"summary":"Получить текущую конфигурацию","description":"Возвращает текущие настройки генерации OTP кодов. Требуется роль ADMIN.","operationId":"getConfig","responses":{"200":{"description":"Конфигурация получена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OtpCodeConfigDTO"}}}},"403":{"description":"Доступ запрещен"}}},"patch":{"tags":["Конфигурация OTP"],"summary":"Изменить конфигурацию","description":"Обновляет настройки генерации OTP кодов. Требуется роль ADMIN.","operationId":"changeConfig","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpCodeConfigDTO"}}},"required":true},"responses":{"200":{"description":"Конфигурация обновлена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OtpCodeConfigDTO"}}}},"400":{"description":"Некорректные параметры","content":{"*/*":{"schema":{"type":"object"}}}},"403":{"description":"Доступ запрещен"}}}},"/api/user":{"get":{"tags":["Пользователи"],"summary":"Получить информацию о текущем пользователе","description":"Возвращает данные аутентифицированного пользователя. Требуется авторизация","operationId":"userInfo","responses":{"200":{"description":"Данные пользователя","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"403":{"description":"Требуется авторизация","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/user/all":{"get":{"tags":["Пользователи"],"summary":"Получить всех пользователей","description":"Возвращает список всех пользователей. Требуется роль ADMIN.","operationId":"getAllUsers","responses":{"200":{"description":"Список пользователей","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}},"403":{"description":"Доступ запрещен","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/otp/validate":{"get":{"tags":["OTP коды"],"summary":"Проверить OTP код","description":"Валидирует введенный пользователем OTP код. Требуется авторизация","operationId":"validateOtp","parameters":[{"name":"operationId","in":"query","required":true,"schema":{"type":"string"}},{"name":"code","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Код верный","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpValidationResponse"},"examples":{"Success":{"summary":"Успешная валидация","description":"Success","value":{"valid":true,"status":"VERIFIED"}}}}}},"400":{"description":"Код неверный или устарел","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpValidationResponse"},"examples":{"Expired":{"summary":"Код устарел","description":"Expired","value":{"valid":false,"status":"EXPIRED"}},"Used":{"summary":"Код уже использован","description":"Used","value":{"valid":false,"status":"USED"}},"Invalid":{"summary":"Неверный код","description":"Invalid","value":{"valid":false,"status":"INVALID"}},"Notfound":{"summary":"Операция с данным id не найдена","description":"Notfound","value":{"valid":false,"status":"NOT_FOUND"}}}}}},"403":{"description":"Требуется авторизация"}}}},"/api/user/{username}":{"delete":{"tags":["Пользователи"],"summary":"Удалить пользователя","description":"Удаляет пользователя по username. Требуется роль ADMIN.","operationId":"deleteUser","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Пользователь удален","content":{"*/*":{"schema":{"type":"string","example":"User deleted successfully"}}}},"403":{"description":"Доступ запрещен","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Пользователь не найден","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"OperationDTO":{"type":"object","description":"Данные операции для генерации OTP","properties":{"operationId":{"type":"string","description":"Уникальный идентификатор операции","example":"login-123","maxLength":100,"minLength":1},"email":{"type":"string","description":"Email для отправки кода","example":"user@example.com"},"phone":{"type":"string","description":"Телефон для отправки кода","example":"+79001234567","pattern":"^\\+?[0-9]{10,15}$"},"tgUsername":{"type":"string","description":"Telegram username для отправки кода, пользователь должен стартануть бота","example":"username","maxLength":100,"minLength":1}},"required":["operationId"]},"OtpGenerationResponse":{"type":"object","properties":{"code":{"type":"string","description":"Сгенерированный код","example":123456},"status":{"type":"string","description":"Статус доставки","enum":["SENT","PARTIAL","FAILED"],"example":"SENT"},"successfulChannels":{"type":"array","description":"Список успешных каналов доставки","example":["SMS","EMAIL"],"items":{"type":"string"}},"failedChannels":{"type":"object","additionalProperties":{"type":"string"},"description":"Ошибки доставки по каналам (канал: описание ошибки)","example":{"TELEGRAM":"Bot blocked by user"}},"expiresAt":{"type":"string","format":"date-time","description":"Время истечения срока действия кода","example":"2023-12-31T23:59:59"}}},"AuthenticationDTO":{"type":"object","description":"Данные для аутентификации","properties":{"username":{"type":"string","description":"Имя пользователя","example":"user123","maxLength":50,"minLength":2},"password":{"type":"string","description":"Пароль","example":"securePassword123!","maxLength":50,"minLength":8}},"required":["password","username"]},"MessageTemplateDTO":{"type":"object","description":"Шаблон сообщения с OTP кодом","properties":{"template":{"type":"string","description":"Текст шаблона, должен содержать {code}","example":"Ваш код: {code}","minLength":1,"pattern":".*\\{code\\}.*"}},"required":["template"]},"OtpCodeConfigDTO":{"type":"object","description":"Настройки генерации OTP кодов","properties":{"lifetimeInSeconds":{"type":"integer","format":"int64","description":"Время жизни кода в секундах","example":300,"maximum":86400,"minimum":10},"length":{"type":"integer","format":"int32","description":"Длина кода","example":6,"maximum":8,"minimum":4}}},"UserDTO":{"type":"object","description":"DTO для представления данных пользователя","properties":{"username":{"type":"string","description":"Уникальное имя пользователя","example":"john_doe","maxLength":50,"minLength":3},"role":{"type":"string","description":"Роль пользователя в системе","enum":["ROLE_USER","ROLE_ADMIN"],"example":"ROLE_USER"}},"required":["role","username"]},"OtpValidationResponse":{"type":"object","description":"Ответ на запрос валидации OTP кода","properties":{"valid":{"type":"boolean","description":"Результат валидации OTP кода.\nЕсли valid=true - код верный,\nесли valid=false - код неверный или устарел","enum":["true","false"],"example":true},"status":{"type":"string","description":"Статус проверки","enum":["VERIFIED","EXPIRED","USED","INVALID","NOT_FOUND"],"example":"VERIFIED"}}}}}}